+++ src/pages/Panel/Ranges/Ranges.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+interface Props {}
+
+const Ranges: React.FC<Props> = (props) => {
+  return <div>ranges</div>;
+};
+
+export default Ranges;
