+++ src/pages/Panel/Ranges/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Ranges";
