+++ src/pages/Panel/Search/Search.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+interface Props {}
+
+const Search: React.FC<Props> = (props) => {
+  return <div>search</div>;
+};
+
+export default Search;
