+++ src/pages/Panel/Users/style.css
@@ -1,0 +1,4 @@
+.container-users {
+  overflow: auto;
+  height: calc(100vh - 120px);
+}
